AngularJS Theme picker/file adder
This assignment has two parts. Part one is a theme picker. Choosing one
of four radio buttons will select a color theme for the page. Part two
is a file adder. Select one of three existing folders in which to
place, name the file and add it to that folder. Error messages if
something is missing.